requirements.txt
vanna
streamlit
python-dotenv
pandas
openpyxl
sqlalchemy
weaviate-client
fastembed
google-generativeai
